+++ services/payments/src/rabbitmq-setup.service.ts
@@ -11,9 +11,13 @@ export class RabbitMqSetupService {
       'user.created',
       'user.updated',
       'user.deleted',
+      'event.created', // TODO: remove and get title from REST endpoint
       'event.updated',
+      'event.published',
+      'event.completed',
+      'event.cancelled',
       'event.deleted',
-    ]; //TODO: received messages
+    ];
 
     await channel.assertExchange(exchange, 'topic', {
       durable: true,