+++ services/payments/src/tickets/domain/events/ticket-type-created.event.ts
@@ -1,0 +1,23 @@
+import { EventEnvelope } from '../../../commons/domain/events/event-envelope';
+export interface TicketTypeCreatedPayload {
+  eventId: string;
+  ticketTypeId: string;
+  price: number;
+}
+
+export class TicketTypeCreatedEvent implements EventEnvelope<TicketTypeCreatedPayload> {
+  public readonly eventType = 'ticket-type.created';
+  public readonly occurredAt: Date;
+
+  constructor(public readonly payload: TicketTypeCreatedPayload) {
+    this.occurredAt = new Date();
+  }
+
+  toJSON() {
+    return {
+      eventType: this.eventType,
+      occurredAt: this.occurredAt,
+      payload: this.payload,
+    };
+  }
+}