+++ services/payments/src/tickets/domain/repositories/event.repository.interface.ts
@@ -10,11 +10,13 @@ export interface EventRepository {
   update(event: {
     eventId: EventId;
     date?: Date;
+    title?: string;
     status: EventStatus;
   }): Promise<Event>;
   updateStatus(eventId: EventId, status: EventStatus);
   delete(id: string): Promise<void>;
   deleteAll(): Promise<void>;
+  isDuplicateError(error: unknown): boolean;
 }
 
 export const EVENT_REPOSITORY = Symbol('EVENT_REPOSITORY');